+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog for EtwinBar
 This changelog is formatted as recommended by [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## [Unreleased]
+### Modified
+- [Config] The "hiddenBlocks" list is replaced by a "blocksVisibility" object. The visibility is set by a 0/1 value and no more adding/removing keys.
+
 ## [12.0] - 2025-08-11
 ### Added
 - [Config] You can totally remove a link by setting it to "null" in the config file